+++ src/Workspaces/CSharp/Portable/CodeGeneration/CSharpCodeGenerationOptions.cs
@@ -4,33 +4,82 @@
 
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.AddImport;
 using Microsoft.CodeAnalysis.CodeGeneration;
+using Microsoft.CodeAnalysis.CodeStyle;
+using Microsoft.CodeAnalysis.CSharp.CodeStyle;
+using Microsoft.CodeAnalysis.Editing;
+using Microsoft.CodeAnalysis.Options;
+using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.CodeGeneration
 {
-    internal sealed class CSharpCodeGenerationContextInfo : CodeGenerationContextInfo
+    internal sealed class CSharpCodeGenerationOptions : CodeGenerationOptions
     {
-        private readonly CSharpCodeGenerationOptions _options;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedMethods;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedAccessors;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedProperties;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedIndexers;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedConstructors;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedOperators;
+        public readonly ExpressionBodyPreference PreferExpressionBodiedLocalFunctions;
+        public readonly NamespaceDeclarationPreference NamespaceDeclarations;
+        public readonly AddImportPlacement PreferredUsingDirectivePlacement;
+        public readonly LanguageVersion LanguageVersion;
 
-        public CSharpCodeGenerationContextInfo(CodeGenerationContext Context, CSharpCodeGenerationOptions Options)
-            : base(Context)
+        public CSharpCodeGenerationOptions(
+            bool placeSystemNamespaceFirst,
+            ExpressionBodyPreference preferExpressionBodiedMethods,
+            ExpressionBodyPreference preferExpressionBodiedAccessors,
+            ExpressionBodyPreference preferExpressionBodiedProperties,
+            ExpressionBodyPreference preferExpressionBodiedIndexers,
+            ExpressionBodyPreference preferExpressionBodiedConstructors,
+            ExpressionBodyPreference preferExpressionBodiedOperators,
+            ExpressionBodyPreference preferExpressionBodiedLocalFunctions,
+            NamespaceDeclarationPreference namespaceDeclarations,
+            AddImportPlacement preferredUsingDirectivePlacement,
+            LanguageVersion languageVersion)
+            : base(placeSystemNamespaceFirst)
         {
-            _options = Options;
+            PreferExpressionBodiedMethods = preferExpressionBodiedMethods;
+            PreferExpressionBodiedAccessors = preferExpressionBodiedAccessors;
+            PreferExpressionBodiedProperties = preferExpressionBodiedProperties;
+            PreferExpressionBodiedIndexers = preferExpressionBodiedIndexers;
+            PreferExpressionBodiedConstructors = preferExpressionBodiedConstructors;
+            PreferExpressionBodiedOperators = preferExpressionBodiedOperators;
+            PreferExpressionBodiedLocalFunctions = preferExpressionBodiedLocalFunctions;
+            NamespaceDeclarations = namespaceDeclarations;
+            PreferredUsingDirectivePlacement = preferredUsingDirectivePlacement;
+            LanguageVersion = languageVersion;
         }
 
-        public new CSharpCodeGenerationOptions Options
-            => _options;
+        public override bool PlaceImportsInsideNamespaces
+            => PreferredUsingDirectivePlacement == AddImportPlacement.InsideNamespace;
 
-        protected override CodeGenerationOptions OptionsImpl
-            => _options;
+        public override CodeGenerationContextInfo GetInfo(CodeGenerationContext context)
+            => new CSharpCodeGenerationContextInfo(context, this);
 
-        public new CSharpCodeGenerationContextInfo WithContext(CodeGenerationContext value)
-            => (Context == value) ? this : new(value, Options);
+        public static CSharpCodeGenerationOptions Create(CSharpParseOptions parseOptions, OptionSet documentOptions)
+            => new(
+                placeSystemNamespaceFirst: documentOptions.GetOption(GenerationOptions.PlaceSystemNamespaceFirst, LanguageNames.CSharp),
+                preferExpressionBodiedMethods: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedMethods).Value,
+                preferExpressionBodiedAccessors: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedAccessors).Value,
+                preferExpressionBodiedProperties: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedProperties).Value,
+                preferExpressionBodiedIndexers: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedIndexers).Value,
+                preferExpressionBodiedConstructors: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedConstructors).Value,
+                preferExpressionBodiedOperators: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedOperators).Value,
+                preferExpressionBodiedLocalFunctions: documentOptions.GetOption(CSharpCodeStyleOptions.PreferExpressionBodiedLocalFunctions).Value,
+                namespaceDeclarations: documentOptions.GetOption(CSharpCodeStyleOptions.NamespaceDeclarations).Value,
+                preferredUsingDirectivePlacement: documentOptions.GetOption(CSharpCodeStyleOptions.PreferredUsingDirectivePlacement).Value,
+                languageVersion: parseOptions.LanguageVersion);
 
-        protected override CodeGenerationContextInfo WithContextImpl(CodeGenerationContext value)
-            => WithContext(value);
+        public static new async Task<CSharpCodeGenerationOptions> FromDocumentAsync(Document document, CancellationToken cancellationToken)
+        {
+            var parseOptions = (CSharpParseOptions?)document.Project.ParseOptions;
+            Contract.ThrowIfNull(parseOptions);
 
-        public static new async ValueTask<CSharpCodeGenerationContextInfo> FromDocumentAsync(CodeGenerationContext context, Document document, CancellationToken cancellationToken)
-            => new(context, await CSharpCodeGenerationOptions.FromDocumentAsync(document, cancellationToken).ConfigureAwait(false));
+            var documentOptions = await document.GetOptionsAsync(cancellationToken).ConfigureAwait(false);
+            return Create(parseOptions, documentOptions);
+        }
     }
 }
